+++ .env.template
@@ -4,7 +4,6 @@ SMART_LLM_MODEL="gpt-4"
 FAST_LLM_MODEL="gpt-3.5-turbo"
 GOOGLE_API_KEY=
 CUSTOM_SEARCH_ENGINE_ID=
-RABBITMQ_HOST=localhost
-RABBITMQ_PORT=5672
-RABBITMQ_USER=guest
-RABBITMQ_PASSWORD=guest
+REDIS_HOST=localhost
+REDIS_PORT=6379
+REDIS_PASSWORD=
