--- rabbitmq-config/definitions.json
@@ -1,42 +1,0 @@
-{
-    "rabbit_version": "3.11",
-    "users": [
-      {
-        "name": "guest",
-        "password": "guest",
-        "tags": "administrator"
-      }
-    ],
-    "vhosts": [
-      {
-        "name": "/"
-      }
-    ],
-    "permissions": [
-      {
-        "user": "guest",
-        "vhost": "/",
-        "configure": ".*",
-        "write": ".*",
-        "read": ".*"
-      }
-    ],
-    "queues": [
-      {
-        "name": "touser",
-        "vhost": "/",
-        "durable": true,
-        "auto_delete": false,
-        "arguments": {
-          "x-message-ttl": 3600000
-        }
-      },
-      {
-        "name": "togpt",
-        "vhost": "/",
-        "durable": true,
-        "auto_delete": false,
-        "arguments": {}
-      }
-    ]
-  }
