--- rabbitmq-config/rabbitmq.conf
@@ -1,4 +1,0 @@
-loopback_users.guest = true
-listeners.tcp.default = 5672
-management.listener.port = 15672
-management.load_definitions = /etc/rabbitmq/definitions.json
