+++ scripts/data/prompt.txt.mustache
@@ -23,7 +23,7 @@ COMMANDS:
 * Write Tests: "write_tests", args: "code": "<full_code_string>", "focus": "<list_of_focus_areas>"
 * Execute Python File: "execute_python_file", args: "file": "<file>"
 * Task Complete (Shutdown): "task_complete", args: "reason": "<reason>"
-{{QA_PROMPTS}}
+{{{QA_PROMPT}}}
 
 RESOURCES:
