+++ scripts/qa/__init__.py
@@ -1,0 +1,10 @@
+import os
+import redis
+
+def connect_to_redis() -> redis.StrictRedis:
+    redis_host = os.environ.get('REDIS_HOST', 'localhost')
+    redis_port = os.environ.get('REDIS_PORT', 6379)
+    redis_password = os.environ.get('REDIS_PASSWORD', None)
+    redis_conn = redis.StrictRedis(host=redis_host, port=redis_port, password=redis_password, decode_responses=True)
+    redis_conn.ping()
+    return redis_conn
